[Horus] Fixes #4205

+++ radio/src/targets/sky9x/board.h
@@ -376,7 +376,7 @@ void eepromStartWrite(uint8_t * buffer, size_t address, size_t size);
 void debugPutc(const char c);
 
 // Telemetry driver
-void telemetryPortInit(uint32_t baudrate, int mode);
+void telemetryPortInit(uint32_t baudrate, uint8_t mode);
 uint32_t telemetryTransmitPending();
 void telemetryTransmitBuffer(uint8_t * buffer, uint32_t size);
 void rxPdcUsart( void (*pChProcess)(uint8_t x) );
